+++ src/error.ts
@@ -76,7 +76,7 @@ export class ArangoError extends ExtendableError {
     const err = new Error(this.message);
     err.name = this.name;
     for (const key of nativeErrorKeys) {
-      if (err[key]) this[key] = err[key];
+      if (err[key]) this[key] = err[key]!;
     }
   }
 }
@@ -95,7 +95,7 @@ export class HttpError extends ExtendableError {
     const err = new Error(this.message);
     err.name = this.name;
     for (const key of nativeErrorKeys) {
-      if (err[key]) this[key] = err[key];
+      if (err[key]) this[key] = err[key]!;
     }
   }
 }