+++ src/test/10-manipulating-collections.ts
@@ -2,17 +2,17 @@ import { expect } from "chai";
 import { Database } from "../arangojs";
 import { DocumentCollection } from "../collection";
 
-const ARANGO_VERSION = Number(process.env.ARANGO_VERSION || 30400);
+const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
+const ARANGO_VERSION = Number(
+  process.env.ARANGO_VERSION || process.env.ARANGOJS_DEVEL_VERSION || 30400
+);
 
 describe("Manipulating collections", function() {
   const name = `testdb_${Date.now()}`;
   let db: Database;
   let collection: DocumentCollection;
   before(async () => {
-    db = new Database({
-      url: process.env.TEST_ARANGODB_URL || "http://localhost:8529",
-      arangoVersion: ARANGO_VERSION
-    });
+    db = new Database({ url: ARANGO_URL, arangoVersion: ARANGO_VERSION });
     await db.createDatabase(name);
     db.useDatabase(name);
   });