+++ src/test/14-document-collections.ts
@@ -2,7 +2,10 @@ import { expect } from "chai";
 import { Database } from "../arangojs";
 import { DocumentCollection } from "../collection";
 
-const ARANGO_VERSION = Number(process.env.ARANGO_VERSION || 30400);
+const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
+const ARANGO_VERSION = Number(
+  process.env.ARANGO_VERSION || process.env.ARANGOJS_DEVEL_VERSION || 30400
+);
 const it3x = ARANGO_VERSION >= 30000 ? it : it.skip;
 
 describe("DocumentCollection API", function() {
@@ -10,10 +13,7 @@ describe("DocumentCollection API", function() {
   let db: Database;
   let collection: DocumentCollection;
   before(async () => {
-    db = new Database({
-      url: process.env.TEST_ARANGODB_URL || "http://localhost:8529",
-      arangoVersion: ARANGO_VERSION
-    });
+    db = new Database({ url: ARANGO_URL, arangoVersion: ARANGO_VERSION });
     await db.createDatabase(name);
     db.useDatabase(name);
   });