+++ src/test/18-graph-edges.ts
@@ -2,16 +2,18 @@ import { expect } from "chai";
 import { Database } from "../arangojs";
 import { Graph } from "../graph";
 
+const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
+const ARANGO_VERSION = Number(
+  process.env.ARANGO_VERSION || process.env.ARANGOJS_DEVEL_VERSION || 30400
+);
+
 describe("Manipulating graph edges", function() {
   const dbName = `testdb_${Date.now()}`;
   const graphName = `testgraph_${Date.now()}`;
   let db: Database;
   let graph: Graph;
   before(async () => {
-    db = new Database({
-      url: process.env.TEST_ARANGODB_URL || "http://localhost:8529",
-      arangoVersion: Number(process.env.ARANGO_VERSION || 30400)
-    });
+    db = new Database({ url: ARANGO_URL, arangoVersion: ARANGO_VERSION });
     await db.createDatabase(dbName);
     db.useDatabase(dbName);
   });