+++ src/test/20-graph-edge-collections.ts
@@ -2,15 +2,17 @@ import { expect } from "chai";
 import { Database } from "../arangojs";
 import { GraphEdgeCollection } from "../graph";
 
+const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
+const ARANGO_VERSION = Number(
+  process.env.ARANGO_VERSION || process.env.ARANGOJS_DEVEL_VERSION || 30400
+);
+
 describe("GraphEdgeCollection API", function() {
   const dbName = `testdb_${Date.now()}`;
   let db: Database;
   let collection: GraphEdgeCollection;
   before(async () => {
-    db = new Database({
-      url: process.env.TEST_ARANGODB_URL || "http://localhost:8529",
-      arangoVersion: Number(process.env.ARANGO_VERSION || 30400)
-    });
+    db = new Database({ url: ARANGO_URL, arangoVersion: ARANGO_VERSION });
     await db.createDatabase(dbName);
     db.useDatabase(dbName);
     const graph = db.graph(`testgraph_${Date.now()}`);